hyperliquid_multiagent_bot_integrated/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
pandas
numpy
ta
prophet
jinja2
